Move parser related tests to own file
