+++ CMakeLists.txt
@@ -215,6 +215,7 @@ set(FlatBuffers_Tests_SRCS
   src/idl_gen_fbs.cpp
   tests/flexbuffers_test.cpp
   tests/monster_test.cpp
+  tests/parser_test.cpp
   tests/test.cpp
   tests/test_assert.h
   tests/test_assert.cpp
